Big refactor of Block.tsx. Also make styles look good.
