+++ frontend/src/components/core/Block/index.tsx
@@ -15,4 +15,6 @@
  * limitations under the License.
  */
 
+// Getting a weird "Default not found in Block" error...
+// eslint-disable-next-line
 export { default } from "./Block"
