+++ frontend/src/components/widgets/Form/styled-components.ts
@@ -17,19 +17,14 @@
 
 import styled from "@emotion/styled"
 
-export interface StyledFormProps {
-  width: number
-}
-
 export const StyledFormSubmitContent = styled.div(() => ({
   display: "flex",
 }))
 
-export const StyledForm = styled.div<StyledFormProps>(({ width, theme }) => ({
+export const StyledForm = styled.div(({ theme }) => ({
   border: `1px solid ${theme.colors.fadedText10}`,
   borderRadius: theme.radii.md,
   padding: "calc(1em - 1px)", // 1px to account for border.
-  width,
 }))
 
 export const StyledErrorContainer = styled.div(({ theme }) => ({
